accessible/tests/mochitest/attributes/test_xml-roles.html

<!DOCTYPE html>
<html>
<head>
  <title>XML landmark tests</title>
  <link rel="stylesheet" type="text/css"
        href="chrome://mochikit/content/tests/SimpleTest/test.css" />

  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <script type="application/javascript"
          src="../common.js"></script>
  <script type="application/javascript"
          src="../role.js"></script>
  <script type="application/javascript"
          src="../attributes.js"></script>

  <script type="application/javascript">

    function doTest()
    {
      // Some AT may look for this
      testAttrs("nav", {"xml-roles" : "navigation"}, true);
      testAttrs("footer", {"xml-roles" : "contentinfo"}, true);
      testAttrs("aside", {"xml-roles" : "complementary"}, true);
      testAttrs("section", {"xml-roles" : "region"}, true);
      testAttrs("main", {"xml-roles" : "main"}, true); // // ARIA override
      testAttrs("form", {"xml-roles" : "form"}, true);

      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTest);
  </script>
</head>
<body>

  <a target="_blank"
     title="Provide mappings for html5 <nav> <header> <footer> <article>"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=593368">
    Bug 593368
  </a><br/>
  <a target="_blank"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=613502"
     title="Map <article> like we do aria role article">
    Bug 613502
  </a>
  <a target="_blank"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=610650"
     title="Change implementation of HTML5 landmark elements to conform">
    Bug 610650
  </a>
  <a target="_blank"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=614310"
     title="Map section to pane (like role=region)">
    Mozilla Bug 614310
  </a>
  <a target="_blank"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=734982"
     title="Map ARIA role FORM">
    Bug 734982
  </a>
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test">
  </pre>

  <nav id="nav">a nav</nav>
  <footer id="footer">a footer</footer>
  <aside id="aside">by the way I am an aside</aside>
  <section id="section">a section</section>
  <article id="main" role="main">a main area</article>
  <article id="form" role="form">a form area</article>

</body>
</html>
